// apps/web/src/services/models/GetWithFriendsGameStats200ResponseInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GeoChess Server API
 * API specification for the GeoChess server.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface GetWithFriendsGameStats200ResponseInner
 */
export interface GetWithFriendsGameStats200ResponseInner {
  /**
   *
   * @type {string}
   * @memberof GetWithFriendsGameStats200ResponseInner
   */
  displayName: string;
  /**
   *
   * @type {number}
   * @memberof GetWithFriendsGameStats200ResponseInner
   */
  wins: number;
  /**
   *
   * @type {number}
   * @memberof GetWithFriendsGameStats200ResponseInner
   */
  losses: number;
}

/**
 * Check if a given object implements the GetWithFriendsGameStats200ResponseInner interface.
 */
export function instanceOfGetWithFriendsGameStats200ResponseInner(
  value: object,
): value is GetWithFriendsGameStats200ResponseInner {
  if (!("displayName" in value) || value["displayName"] === undefined)
    return false;
  if (!("wins" in value) || value["wins"] === undefined) return false;
  if (!("losses" in value) || value["losses"] === undefined) return false;
  return true;
}

export function GetWithFriendsGameStats200ResponseInnerFromJSON(
  json: any,
): GetWithFriendsGameStats200ResponseInner {
  return GetWithFriendsGameStats200ResponseInnerFromJSONTyped(json, false);
}

export function GetWithFriendsGameStats200ResponseInnerFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): GetWithFriendsGameStats200ResponseInner {
  if (json == null) {
    return json;
  }
  return {
    displayName: json["displayName"],
    wins: json["wins"],
    losses: json["losses"],
  };
}

export function GetWithFriendsGameStats200ResponseInnerToJSON(
  json: any,
): GetWithFriendsGameStats200ResponseInner {
  return GetWithFriendsGameStats200ResponseInnerToJSONTyped(json, false);
}

export function GetWithFriendsGameStats200ResponseInnerToJSONTyped(
  value?: GetWithFriendsGameStats200ResponseInner | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    displayName: value["displayName"],
    wins: value["wins"],
    losses: value["losses"],
  };
}