// apps/web/src/services/models/GetWithFriendsGames200ResponseInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GeoChess Server API
 * API specification for the GeoChess server.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface GetWithFriendsGames200ResponseInner
 */
export interface GetWithFriendsGames200ResponseInner {
  /**
   *
   * @type {string}
   * @memberof GetWithFriendsGames200ResponseInner
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof GetWithFriendsGames200ResponseInner
   */
  opponentDisplayName: string;
  /**
   *
   * @type {string}
   * @memberof GetWithFriendsGames200ResponseInner
   */
  result: GetWithFriendsGames200ResponseInnerResultEnum;
  /**
   *
   * @type {boolean}
   * @memberof GetWithFriendsGames200ResponseInner
   */
  expired: boolean;
  /**
   *
   * @type {Date}
   * @memberof GetWithFriendsGames200ResponseInner
   */
  createdAt: Date;
  /**
   *
   * @type {Date}
   * @memberof GetWithFriendsGames200ResponseInner
   */
  updatedAt: Date;
}

/**
 * @export
 */
export const GetWithFriendsGames200ResponseInnerResultEnum = {
  Win: "win",
  Lose: "lose",
} as const;
export type GetWithFriendsGames200ResponseInnerResultEnum =
  (typeof GetWithFriendsGames200ResponseInnerResultEnum)[keyof typeof GetWithFriendsGames200ResponseInnerResultEnum];

/**
 * Check if a given object implements the GetWithFriendsGames200ResponseInner interface.
 */
export function instanceOfGetWithFriendsGames200ResponseInner(
  value: object,
): value is GetWithFriendsGames200ResponseInner {
  if (!("id" in value) || value["id"] === undefined) return false;
  if (
    !("opponentDisplayName" in value) ||
    value["opponentDisplayName"] === undefined
  )
    return false;
  if (!("result" in value) || value["result"] === undefined) return false;
  if (!("expired" in value) || value["expired"] === undefined) return false;
  if (!("createdAt" in value) || value["createdAt"] === undefined) return false;
  if (!("updatedAt" in value) || value["updatedAt"] === undefined) return false;
  return true;
}

export function GetWithFriendsGames200ResponseInnerFromJSON(
  json: any,
): GetWithFriendsGames200ResponseInner {
  return GetWithFriendsGames200ResponseInnerFromJSONTyped(json, false);
}

export function GetWithFriendsGames200ResponseInnerFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): GetWithFriendsGames200ResponseInner {
  if (json == null) {
    return json;
  }
  return {
    id: json["id"],
    opponentDisplayName: json["opponentDisplayName"],
    result: json["result"],
    expired: json["expired"],
    createdAt: new Date(json["createdAt"]),
    updatedAt: new Date(json["updatedAt"]),
  };
}

export function GetWithFriendsGames200ResponseInnerToJSON(
  json: any,
): GetWithFriendsGames200ResponseInner {
  return GetWithFriendsGames200ResponseInnerToJSONTyped(json, false);
}

export function GetWithFriendsGames200ResponseInnerToJSONTyped(
  value?: GetWithFriendsGames200ResponseInner | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    id: value["id"],
    opponentDisplayName: value["opponentDisplayName"],
    result: value["result"],
    expired: value["expired"],
    createdAt: value["createdAt"].toISOString(),
    updatedAt: value["updatedAt"].toISOString(),
  };
}